+++ CHANGELOG.md
@@ -4,6 +4,8 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Unreleased
 
+- Move `django.startapp` command from `cookiecutter` to `copier`
+
 ## 1.8.0
 
 - Add commands for `uv`